# docs/swagger.yaml
basePath: /
definitions:
  domain.Campaign:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      user_id:
        type: string
    type: object
...
  domain.HierarchyNode:
    properties:
      children:
        items:
          $ref: '#/definitions/domain.HierarchyNode'
        type: array
      name:
        type: string
      type:
        description: '"root", "campaign", "link"'
        type: string
      value:
        description: E.g., clicks (optional)
        type: integer
    type: object
  domain.Link:
    properties:
      campaign_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      slug:
        type: string
      target_url:
        type: string
      user_id:
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  v1.CampaignResponse:
    properties:
      data:
        $ref: '#/definitions/domain.Campaign'
    type: object
  v1.CampaignsListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.Campaign'
        type: array
    type: object
  v1.CreateCampaignRequest:
    properties:
      name:
        example: My Campaign
        type: string
    required:
    - name
    type: object
  v1.CreateLinkRequest:
    properties:
      campaign_id:
        type: string
      target_url:
        example: https://google.com
        type: string
    required:
    - target_url
    type: object
...
  v1.LinkResponse:
    properties:
      data:
        $ref: '#/definitions/domain.Link'
    type: object
  v1.LinksListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.Link'
        type: array
    type: object
...
  v1.ProfileTreeResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/domain.HierarchyNode'
        description: Предпологаем, что HierarchyNode определен в domain
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Traffic analytics service
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Traffic Analytics
  version: "1.0"
paths:
...
  /api/v1/campaigns:
    get:
      description: Get list of user campaigns
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.CampaignsListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get campaigns
      tags:
      - Campaigns
    post:
      consumes:
      - application/json
      description: Create a new campaign
      parameters:
      - description: Campaign Name
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.CreateCampaignRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v1.CampaignResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create campaign
      tags:
      - Campaigns
  /api/v1/links:
    get:
      description: Get links with filtering, search and pagination
      parameters:
      - description: Filter by Campaign ID
        in: query
        name: campaign_id
        type: string
      - description: Search by slug or target URL
        in: query
        name: search
        type: string
      - description: Filter by active status
        in: query
        name: is_active
        type: boolean
      - description: Limit (default 10)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.LinksListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get links list
      tags:
      - Links
    post:
      consumes:
      - application/json
      description: Shorten a URL
      parameters:
      - description: Link Info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.CreateLinkRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v1.LinkResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Limit reached or Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create link
      tags:
      - Links
  /api/v1/links/{id}:
    delete:
      description: Remove a link
      parameters:
      - description: Link ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete link
      tags:
      - Links
  /api/v1/users/tree:
    get:
      description: Get hierarchical structure of User -> Campaigns -> Links for Tree
        visualization
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ProfileTreeResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Profile Hierarchy
      tags:
      - Users
...
securityDefinitions:
  APIKeyAuth:
    description: Static API Key for admin/service access
    in: header
    name: X-API-Key
    type: apiKey
  BasicAuth:
    type: basic
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"